subscriber/requirements.txt
paho-mqtt>=2.0.0
flask>=2.0.0
